Cleanup, move "public" methods to main module.
