+++ docs/HISTORY.txt
@@ -6,4 +6,4 @@ Changelog
 ------------------
 
 - Initial implementation.
-  [lgraf]
+  [lgraf, deif]
