+++ ooxml_docprops/properties.py
@@ -256,20 +256,3 @@ class OOXMLDocument(OOXMLPackage):
     def update_properties(self, metadata):
         assert not self._read_only, 'you may not update readonly documents!'
         self.properties = self.properties.update(metadata)
-
-
-def update_properties(document, metadata):
-    """Update custom doc properties in the document specified by path
-    `document` with properties from `metadata`. Modifies the document in place!
-    """
-    with OOXMLDocument(document) as doc:
-        doc.update_properties(metadata)
-
-
-def read_properties(document):
-    """Read custom doc properties from the file `document`.
-    """
-    with OOXMLDocument(document, read_only=True) as doc:
-        for name in doc.properties.get_property_names():
-            value = doc.properties.get_property_value(name)
-            yield (name, value)
